refactor: implement vendor overview charts and forms, reorganize vendor registration components
